requirements.txt
# Core NLP
spacy>=3.7.0
sentence-transformers>=2.6.0
scikit-learn>=1.4.0

# PDF & Document parsing
pdfplumber>=0.11.0
python-docx>=1.1.0
PyMuPDF>=1.24.0

# Data
numpy>=1.26.0
pandas>=2.2.0

# UI
streamlit>=1.35.0

# Utilities
regex>=2024.4.16